# services/main-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Central API for SalesAI Intelligent Revenue Intelligence \u0026 Coaching SaaS",
        "title": "SalesAI API",
        "contact": {
            "name": "API Support",
            "email": "support@salesai.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/calls": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of calls for the company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "List calls with filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by manager ID",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/calls.ListCallsResponse"
                        }
                    }
                }
            }
        },
        "/calls/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a call, including transcript and analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "Get call details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Call ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/calls/{id}/analysis": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the AI-generated analysis report for a call",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "Get call analysis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Call ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.AnalysisReport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/calls/{id}/audio": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stream the audio recording of a call from MinIO",
                "produces": [
                    "audio/mpeg"
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "Stream call audio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Call ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/calls/{id}/reprocess": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually trigger STT and AI analysis for a failed or old call",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "Re-queue call for processing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Call ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/calls/{id}/transcript": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the diarized transcript of a call",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "Get call transcript",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Call ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Transcript"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/companies/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get preferences and settings for the company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Get company settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Company"
                        }
                    }
                }
            }
        },
        "/companies/{id}/settings": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update STT and LLM provider preferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Update company preferences",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Settings Request (stt_model_preference, llm_provider)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "calls.ListCallsResponse": {
            "type": "object",
            "properties": {
                "calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Call"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "domain.AnalysisReport": {
            "type": "object",
            "properties": {
                "brief": {
                    "type": "string"
                },
                "call_id": {
                    "type": "string"
                },
                "errors_free": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "kpi": {
                    "type": "number"
                },
                "llm_provider": {
                    "type": "string"
                },
                "next_best_action": {
                    "type": "string"
                },
                "overall_rating": {
                    "type": "number"
                },
                "processed_at": {
                    "type": "string"
                },
                "quality_score": {
                    "type": "integer"
                },
                "recommendation": {
                    "type": "string"
                },
                "script_id": {
                    "type": "string"
                },
                "script_match": {
                    "type": "integer"
                }
            }
        },
        "domain.Call": {
            "type": "object",
            "properties": {
                "call_date": {
                    "type": "string"
                },
                "call_link": {
                    "type": "string"
                },
                "call_time": {
                    "type": "string"
                },
                "chat_link": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string"
                },
                "client_phone": {
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "manager_id": {
                    "type": "string"
                },
                "manager_name": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.CallStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.CallStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "completed",
                "error"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusProcessing",
                "StatusCompleted",
                "StatusError"
            ]
        },
        "domain.Company": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "llm_provider": {
                    "$ref": "#/definitions/domain.LLMProvider"
                },
                "name": {
                    "type": "string"
                },
                "stt_model_preference": {
                    "$ref": "#/definitions/domain.STTModel"
                },
                "subscription_tier": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.LLMProvider": {
            "type": "string",
            "enum": [
                "openai",
                "gemini"
            ],
            "x-enum-varnames": [
                "LLMOpenAI",
                "LLMGemini"
            ]
        },
        "domain.STTModel": {
            "type": "string",
            "enum": [
                "whisperx_local",
                "openai",
                "gemini"
            ],
            "x-enum-varnames": [
                "STTWhisperXLocal",
                "STTOpenAI",
                "STTGemini"
            ]
        },
...
        "domain.Transcript": {
            "type": "object",
            "properties": {
                "call_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "processed_at": {
                    "type": "string"
                },
                "processing_time_seconds": {
                    "type": "integer"
                },
                "speaker_diarized_json": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TranscriptSegment"
                    }
                },
                "stt_provider": {
                    "type": "string"
                }
            }
        },
        "domain.TranscriptSegment": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "number"
                },
                "speaker": {
                    "type": "string"
                },
                "start": {
                    "type": "number"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}